restructured test cases
