+++ test/listener_test.py
@@ -19,12 +19,12 @@ logging.getLogger('').setLevel(logging.DEBUG)
 class ListenerImportTest(unittest.TestCase):
 
     def testClassesImportable(self):
-        assert hasattr(can, 'Listener')
-        assert hasattr(can, 'BufferedReader')
-        assert hasattr(can, 'Notifier')
-        assert hasattr(can, 'ASCWriter')
-        assert hasattr(can, 'CanutilsLogWriter')
-        assert hasattr(can, 'SqlReader')
+        self.assertTrue(hasattr(can, 'Listener'))
+        self.assertTrue(hasattr(can, 'BufferedReader'))
+        self.assertTrue(hasattr(can, 'Notifier'))
+        self.assertTrue(hasattr(can, 'ASCWriter'))
+        self.assertTrue(hasattr(can, 'CanutilsLogWriter'))
+        self.assertTrue(hasattr(can, 'SqlReader'))
         # TODO add more?
 
 
@@ -63,51 +63,6 @@ class ListenerTest(BusTest):
         a_listener(generate_message(0xDADADA))
         m = a_listener.get_message(0.2)
         self.assertIsNotNone(m)
-
-    def testSQLWriterWritesToSameFile(self):
-        f = tempfile.NamedTemporaryFile('w', delete=False)
-        f.close()
-
-        first_listener = can.SqliteWriter(f.name)
-        first_listener(generate_message(0x01))
-
-        sleep(first_listener.MAX_TIME_BETWEEN_WRITES)
-        first_listener.stop()
-
-        second_listener = can.SqliteWriter(f.name)
-        second_listener(generate_message(0x02))
-
-        sleep(second_listener.MAX_TIME_BETWEEN_WRITES)
-
-        second_listener.stop()
-
-        con = sqlite3.connect(f.name)
-
-        with con:
-            c = con.cursor()
-
-            c.execute("select COUNT() from messages")
-            self.assertEqual(2, c.fetchone()[0])
-
-            c.execute("select * from messages")
-            msg1 = c.fetchone()
-            msg2 = c.fetchone()
-
-        self.assertEqual(msg1[1], 0x01)
-        self.assertEqual(msg2[1], 0x02)
-
-
-class BLFTest(unittest.TestCase):
-
-    def test_reader(self):
-        logfile = os.path.join(os.path.dirname(__file__), "data", "logfile.blf")
-        messages = list(can.BLFReader(logfile))
-        self.assertEqual(len(messages), 1)
-        self.assertEqual(messages[0],
-                         can.Message(
-                             extended_id=False,
-                             arbitration_id=0x64,
-                             data=[0x1, 0x2, 0x3, 0x4, 0x5, 0x6, 0x7, 0x8]))
 
 
 if __name__ == '__main__':
