+++ test/logformats_test.py
@@ -11,6 +11,8 @@ again.
 import unittest
 import tempfile
 from time import sleep
+import sqlite3
+import os
 
 try:
     # Python 3
@@ -22,7 +24,8 @@ except ImportError:
 import can
 
 from data.example_data import TEST_MESSAGES_BASE, TEST_MESSAGES_REMOTE_FRAMES, \
-                              TEST_MESSAGES_ERROR_FRAMES, TEST_COMMENTS
+                              TEST_MESSAGES_ERROR_FRAMES, TEST_COMMENTS, \
+                              generate_message
 
 def _test_writer_and_reader(test_case, writer_constructor, reader_constructor, sleep_time=None,
                             check_remote_frames=True, check_error_frames=True,
@@ -75,15 +78,18 @@ def _test_writer_and_reader(test_case, writer_constructor, reader_constructor, s
         # write messages and insert comments here and there
         # Note: we make no assumptions about the length of original_messages and original_comments
         for msg, comment in zip_longest(original_messages, original_comments, fillvalue=None):
-            # msg and comment might ne None
+            # msg and comment might be None
             if comment is not None:
                 writer.log_event(comment) # we already know that this method exists
+                print("writing comment: ", comment)
             if msg is not None:
                 writer(msg)
+                print("writing message: ", msg)
     else:
         # ony write messages
         for msg in original_messages:
             writer(msg)
+            print("writing message: ", msg)
 
     # sleep and close the writer
     if sleep_time is not None:
@@ -143,6 +149,38 @@ class TestSqlFileFormat(unittest.TestCase):
                                 sleep_time=0.5,
                                 check_comments=False)
 
+    def testSQLWriterWritesToSameFile(self):
+        f = tempfile.NamedTemporaryFile('w', delete=False)
+        f.close()
+
+        first_listener = can.SqliteWriter(f.name)
+        first_listener(generate_message(0x01))
+
+        sleep(first_listener.MAX_TIME_BETWEEN_WRITES)
+        first_listener.stop()
+
+        second_listener = can.SqliteWriter(f.name)
+        second_listener(generate_message(0x02))
+
+        sleep(second_listener.MAX_TIME_BETWEEN_WRITES)
+
+        second_listener.stop()
+
+        con = sqlite3.connect(f.name)
+
+        with con:
+            c = con.cursor()
+
+            c.execute("select COUNT() from messages")
+            self.assertEqual(2, c.fetchone()[0])
+
+            c.execute("select * from messages")
+            msg1 = c.fetchone()
+            msg2 = c.fetchone()
+
+        self.assertEqual(msg1[1], 0x01)
+        self.assertEqual(msg2[1], 0x02)
+
 class TestBlfFileFormat(unittest.TestCase):
     """Tests can.BLFWriter and can.BLFReader"""
 
@@ -150,6 +188,16 @@ class TestBlfFileFormat(unittest.TestCase):
         _test_writer_and_reader(self, can.BLFWriter, can.BLFReader,
                                 sleep_time=None,
                                 check_comments=False)
+
+    def test_reader(self):
+        logfile = os.path.join(os.path.dirname(__file__), "data", "logfile.blf")
+        messages = list(can.BLFReader(logfile))
+        self.assertEqual(len(messages), 1)
+        self.assertEqual(messages[0],
+                         can.Message(
+                             extended_id=False,
+                             arbitration_id=0x64,
+                             data=[0x1, 0x2, 0x3, 0x4, 0x5, 0x6, 0x7, 0x8]))
 
 
 if __name__ == '__main__':
